Move AppSettings to DataModel package
