+++ Modules/DataModel/Sources/PocketCastsDataModel/Public/Model/AppSettings.swift
